Better split stuff in interfaces package
