+++ packages/core/interfaces/src/index.ts
@@ -1,2 +1,4 @@
-export * from './compilerSupport.js'
-export * from './fluenceClient.js'
+export * from './compilerSupport/aquaTypeDefinitions.js';
+export * from './compilerSupport/compilerSupportInterface.js';
+export * from './commonTypes.js';
+export * from './fluenceClient.js';
